pyproject.toml
[project]
name = "customer-churn-databricks"
version = "0.0.1"
description = "MLOps Project with Databricks: Customer Churn Prediction"
keywords = ["MLOps", "Databricks", "Machine Learning", "Customer Churn"]
authors = [{ name = "Your Name"}]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = ["lightgbm>=4.5.0, <5",
                "scikit-learn>=1.5.1, <2",
                "cloudpickle>=3.0.0, <4",
                "mlflow==2.17.2",
                "numpy>=1.26.4, <2",
                "pandas>=2.2.2, <3",
                "pyarrow>=14.0.0, <15",
                "cffi>=1.17.1, <2",
                "scipy>=1.14.1, <2",
                "matplotlib>=3.9.2, <4",
                "databricks-feature-engineering==0.6",
                "databricks-feature-lookup==1.2.0",
                "imbalanced-learn>=0.11.0, <1",
                "loguru>=0.6.0, <1",
                "pre-commit>=2.20.0, <3",
                "pydantic==2.9.0, <3",
                "databricks-sdk==0.32.0",
                ]

[project.optional-dependencies]
dev = ["databricks-connect>=15.4.1, <16",
       "ipykernel>=6.29.5, <7",
       "pip>=24.2",
       "pytest>=8.1.1, <9",
       "ruff>=0.2.0, <1" ,
       "python-dotenv>=1.0.0, <2"
]

# Build system configuration
[build-system]
requires = ["hatchling>=1.23.0"]
build-backend = "hatchling.build"

# Hatch configuration for src layout
[tool.hatch.build]
packages = ["src"]

# Wheel build configuration
[tool.hatch.build.targets.wheel]
packages = ["src/customer_churn"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "F",  # pyflakes rules
    "E",  # pycodestyle error rules
    "W",  # pycodestyle warning rules
    "B",  # flake8-bugbear rules
    "I",  # isort rules
]
ignore = [
    "E501",  # line-too-long
]

[tool.ruff.format]
indent-style = "space"
